[IMP] pos_*: override payment validation dependencies in related modules
In this commit:

- We have overidden `OrderPaymentValidation` class method in related modules.
