+++ addons/l10n_pe_pos/static/src/overrides/components/paymentt_screen/payment_screen.js
@@ -1,5 +1,3 @@
-import { _t } from "@web/core/l10n/translation";
-import { AlertDialog } from "@web/core/confirmation_dialog/confirmation_dialog";
 import { PaymentScreen } from "@point_of_sale/app/screens/payment_screen/payment_screen";
 import { patch } from "@web/core/utils/patch";
 
@@ -9,24 +7,5 @@ patch(PaymentScreen.prototype, {
         if (this.pos.isPeruvianCompany()) {
             this.currentOrder.setToInvoice(true);
         }
-    },
-    async _isOrderValid(isForceValidate) {
-        const res = await super._isOrderValid(...arguments);
-        if (!this.pos.isPeruvianCompany() && res) {
-            return res;
-        }
-        if (!res) {
-            return false;
-        }
-        const currentPartner = this.currentOrder.getPartner();
-        if (currentPartner && !currentPartner.vat) {
-            this.pos.editPartner(currentPartner);
-            this.dialog.add(AlertDialog, {
-                title: _t("Missing Field"),
-                body: _t("An Identification Number Is Required"),
-            });
-            return false;
-        }
-        return res;
     },
 });
